amount int types have been changed to double

+++ Challenge.Persistence/DTOs/BalanceDTO.cs
@@ -4,8 +4,8 @@
     {
         public Guid Id { get; set; }
         public Guid? UserId { get; set; }
-        public long? AvailableBalance { get; set; }
-        public long? BlockedBalance { get; set; }
+        public double? AvailableBalance { get; set; }
+        public double? BlockedBalance { get; set; }
         public string? Currency { get; set; }
         public DateTime? LastUpdated { get; set; }
     }
